feat(si-tmpl): component commands
Adds commands for component get, component update, component delete, and
component search. 

Get and update are symmetric - you can take the yaml output of a get and
feed it back to a delete, and get an update operation that will be
idempotent (but depending on how you write those attribute updates,
maybe not!).

Search supports returning specific attributes or the full component. A
future addition will add the option to have the full component mode have
a symmetric editing partner.