fix(web): Remove the need for the AI Agent banner
When a user onboards into the product, we lock them into the onboarding 
flow until the AI Agent starts.

When they invite a user to the workspace, the user isn’t prompted for 
the onboarding flow IF there are components on HEAD. Components on HEAD 
here means that someone has completed the onboarding for the workspace 
AND the workspace is active.

Those invited users were they continually prompted to start the AI Agent
and couldn’t remove that warning.

This warning is not relevant any more as someone in the workspace is 
already onboarded to the AI Agent