fixes for ComponentContextMenu bugs
How does this PR change the system?
This PR fixes the following bugs related to the ComponentContextMenu -

The menu now closes when the Upgrade option is selected just like any other option
The menu no longer fails to execute certain actions (delete, erase, template) when they are invoked via a hotkey press while the menu is not open.
The menu no longer loses the component ids assigned to a particular action in certain obscure scenarios.
Finishing an action (either via the menu or hotkey) now clears the component selection after actions where doing so should happen.
The menu no longer shows actions which can only be performed on a single component when multiple components are selected.

Out of Scope:
No changes to the menu were made besides fixing these bugs.

How was it tested?
Ran through each menu action using both the menu and hotkeys, and for hotkeys tested with the menu open and closed.