Onboarding V4 - Hetzner & DigitalOcean
How does this PR change the system?
This PR updates the onboarding flow to support Hetzner and DigitalOcean

It also revamps and refactors the onboarding flow code to make adding new providers even simpler

The DigitalOcean option in onboarding is behind a feature flag so we can wait to roll it out until the DigitalOcean assets are ready for users

Out of Scope:
We will be adding additional providers to the onboarding as we support them.

How was it tested?
Ran through the onboarding for each provider successfully!

Double checked all the copy and design.