always force line numbers in CodeViewer
How does this PR change the system?

CodeMirror has a bug where it can't handle a string with exactly 1 line of text with no newlines.

Previously we handled this by not using CodeMirror for such strings in CodeViewer and instead rendering a raw monospaced string in a div. This was a suboptimal solution for many reasons, primarily because it can break in ways that CodeMirror doesn't and has to be maintained separately.

Now the CodeViewer will just add a newline to ensure CodeMirror always works. This feature was previously called forceLineNumbers and is now just always turned on.

Out of Scope:
This PR does not add or change any other features of the CodeViewer

How was it tested?
Checked all existing CodeViewers to make sure none of them broke!