bug fix for componentDiscover
How does this PR change the system?
Fixes a bug in the componentDiscover MCP tool which caused discovery to fail in certain situations, most frequently in a new workspace with a schema that is not installed yet.

A race condition can occur when the template component is created and the tool attempts to gather the default variant data immediately afterwards - this addition to the tool's code prevents the issue by automatically reattempting the call for the default variant data when it detects that waiting is necessary.

This bug could either result in a failure with no data at all (the error message occurring on the variantFunctions.find call) or a failure with an empty array of variantFuncs causing the "The schema doesn't support discovery" error on a schema which does have a discovery function.

Out of Scope:
We may want to change the getDefaultVariant endpoint in the future to avoid the workaround in the MCP tool.

How was it tested?
Ran discovery in a wide range of scenarios successfully!