Add monotonic counters for change set index fetch
This change adds four monotonic counters for change set index fetch
calls. Specifically, it tracks three "None" cases as well as one
error case. This is particularly useful for knowing the frequency at
which certain scenarios for not being able to fetch a change set index
happens. A new grafana dashboard has been added for these counters.

A secondary part of this change is making the error returned by luminork
for any of these scenarios a 424 error instead of a 500.

The genesis of this change was from a situation where luminork kept
returning an error for a missing change set index that was actually
caused by a definition mismatch. It was caused by a newer luminork
version being deployed than the other services, which would permanently
cause luminork to not be able to return a change set index when using
the search API. With this change, that would have been caught much
sooner.

Signed-off-by: Nick Gerace <nick@systeminit.com>