Better Import UI in ComponentDetails
How does this PR change the system?

This PR improves the UI for importing a resource in the `ComponentDetails` page in the following ways -
- Replaced the janky existing form with one which pops up in a modal
- The import status is now more accurately reflected on the import button and the banner
- Keyboard shortcut for opening the import modal (the I key)

Out of Scope:

This PR does not change the format of the import status banner, which we may want to change in the future.

This PR also does not touch the management function section which will likely need to be adjusted as well.

How was it tested?

Ran imports successfully on a number of components, also tested a variety of failure states.