Disable tile-view test until we fix it
This change disables the tile-view test until we can fix. It has been
thrashing in tools for weeks and we have grown numb to the alert. Time
to end the pain since we are not getting value from it currently.

This change also gets rid of leading and trailing whitespaces in the
"e2e-validation" workflows file.

Signed-off-by: Nick Gerace <nick@systeminit.com>