fix(versitygw): update location of binary in image from upstream change
The location of the `versitygw` binary was moved from `/app/versitygw`
to `/usr/local/bin/versitygw` in the upstream image. While this doesn't
affect how the image is built, the invocation of it via our Docker
Compose setups need to be adjusted. This change updates the development
Docker Compose as well as the CI Docker Compose which is used with all
the integration tests.

References: versity/versitygw#1569

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>