Add building schema metrics
This change adds "building schema" metrics for routes in luminork that
can return 202s back to the AI agent. These metrics are useful for
detecting unavailable data for MVs for cached modules as well as the
trends and spikes for when these cases are hit. A corresponding
dashboard has been added for these new metrics.

This change also adjusts the "component-discover" tool response to
prompt the user or agent to try discovery again in case the attempt
count is reached. In addition, the discovery component will be erased
for cleanup when this happens.

Additional changes
- Remove unused "ws.rs" file from "sdf-server"

Signed-off-by: Nick Gerace <nick@systeminit.com>