Link spans to compressed edda request spans
This change links spans from an edda request to its compressed message.
As a result of this change, the compressing stream's span has been
bumped from DEBUG to INFO and is the linking span between core
sdf/luminork/pinga/rebaser traces and the ebba build MV traces. This is
required for being able to eventually trace a user action to the
building of a materialized view and patches sent back to the frontend.

Co-authored-by: John Keiser <jkeiser@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>