Improved Actions panel!
How does this PR change the system?

This PR improves the UI/UX of the ActionQueueList which shows in Explore - the new version better shows the state of the listed actions and which actions are dependent on others. It is also more consistent with the new UI, as the old version was pulled directly from the old UI and never updated.

Out of Scope:

For now, these UI changes are particular to this one place in the UI and other places we're displaying actions still use the old ActionCard - eventually we will probably unify the components and eliminate the old ActionCard.

We also might want to adjust the UI in the future to make it more clear which actions are blocked from running by failed actions - this is not more clear in the new version than it was in the old version.

Also there are docs implications explained below.

How was it tested?

Ran through a wide variety of possible action scenarios with multiple parents and multiple levels of parentage in different states.

Does it require a docs change?

Yes, and this PR does not include it (reasoning below)

We will probably want to update the docs with new screenshots to reflect the new actions panel, but for now the panel is similar enough that users should still be able to use the docs without an update.