Add missing S3->PG fallback metrics
There were a couple of code paths that were not emitting the layer
fallback metrics. Specifically, `get_bytes_from_durable_storage`, and
`get_bulk`. We now record the fallback in both, and `get_bulk`
increments the fallback counter by the number of items that needed to be
fetched from PG (instead of treating the entire `get_bulk` as a single
"fallback").