fixed actions scroll area in ApplyChangeSetModal
How does this PR change the system?
This PR is a small adjustment to the scrolling area that displays actions in the ApplyChangeSet modal so that the header is not part of the scrolling area. This change also fixes the scrollbar in the area to match the styling of other scrollbars in the new UI.

Out of Scope:
This PR does not update the ApplyChangeSetModal actions section to match the new actions panel. We may do that in the future, retaining this improvement when we do.

How was it tested?
This small change required only minor manual testing.

Does it require a docs change?
No