web experience backstop & telemetry
there is an outside chance that the system will fail to send a wsevent
(incident, poor timing of a deploy, etc etc)
to the web worker that tells it the index has been built successfully
so we have a backup cold start interval here
to prevent a user from being stuck in the lobby perpetually

make sure we end a span in a no-op case

bring back a disconnect notification for only SI cohort

emit span events for when we open and close ws connections

every 30 seconds, if we have not gotten index update messages, try to cold start, just in case
this is a fail safe for clients that are not receiving messages.
yes, in a quiet workspace this will make calls that result in nothing much happening.
but its hitting a cache and its a small price to pay.
30 seconds might be the wrong number, but its something.