auth-api: support cli redirect flow
If the `cli_redir` query param is set to a number, we will redirect to
`localhost:{cli_redir}` with a nonce, instead of back to the systeminit
web app, in order to enable the si cli auth flow. The nonce can be used
to acquire the auth-api token generated by the auth flow (and is
discarded after a single use, or after 500 seconds).

This must be in prod first before we can release the si-cli with login
flow.