feat(telemetry): add custom histogram bucket configuration infrastructure
Custom histogram buckets need to be defined as part of the OpenTelemetry
pipeline. In order to make this a bit easier in the future, this can now
be done in a data-centric approach by creating new HistogramBucketConfig
and returning it as part of the Vec in `custom_histogram_configs`,
allowing metrics to use custom buckets without any "code" changes in the
main `otel_metrics` function.

This also sets up custom buckets for `layer_cache_insert_size_bytes`
which will be the first to take advantage of this capability.