Follow spans from edda to websocket publish
This change adds two critical parentage setup calls in order to follow
spans from the moment an MV is built to right before the message is
published on the websocket for client webworkers. This is a follow-up
to PRs #7902 and #7910. Before all this work, the following
relationships were not instrumented:

- Alpha: edda (per mutation request) to edda (compressed)
- Beta: edda (compressed) to sdf (multiplexer)
- Charlie: sdf (multiplexer) to sdf (bifrost)
- Delta: sdf (bifrost) to webworker

PR #7902 instrumented relationship Alpha with span linking. Linking was
required because N edda requests are compressed down to one. Spans can
only have one parent, so linking to N spans was the move.

This change instruments relationships Beta and Charlie. Relationship
Beta helps answer the question, "did my MV build message make it from
edda to sdf?". Relationship Charlie helps answer the question "did all
the internal multiplexing and setup work and are we about to send the
MV build message via websockets?".

This is great and all, but what about the total number of spans
introduced by this change? Fortunately for PR #7902, we only added
one additional span per inbound request to the main edda change set
handler. We were already tracing each handler call at the INFO level and
the additional span count was "low impact"... well, at least that's
what we see so far. However for this change, the impact is a bit
greater. There are now two additional spans per patch batch and there
can be N of those for a given compressed edda request. Essentially,
any message that will be sent to client webworkers related to edda from
sdf will now be instrumented at the INFO level.

There's good news though. For one, I got re-acquainted with my old
multiplexer code. It's not quite 3000 years old, but as I was correcting
the "error!" messages, it sure felt like it. Pride aside, the changes
involving the multiplexer and its payloads were minimal and only
affect edda in sdf. All other usages of the multiplexer intentionally
are not instrumented in order to keep the impact on total number of
spans minimal. That even includes skipping the span within edda that
sends out the messages to sdf, which intentionally remains at a DEBUG
level.

As a consequence of these changes, there is now a
"MultiplexerRequestPayload" type that contains both the NATS message
and an optional OpenTelemetryContext. The latter is only populated
and used for relationships Beta and Charlie, but could prove useful
for other multiplexer clients in the future. This type is helpful
anyway as users of the "nats-multiplexer-client" crate now do not have
to import the "Message" type from "si-data-nats" directly.

All of those should prove useful in determining what patches are seen
in the webworker after a given mutation(e.g. updating a component's
name) to the system. That's the hope. WE RIDE.

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>