Ensure last leaf binding removal keeps one proto
This change ensures that when the last leaf binding is removed, an
attribute prototype for the "unset" intrinsic func is created. This
reflects the default state of the variant's leaf setup. Without this
change, iteratively working on schemas with the authoring subsystem
could result in schema variant MV build failures.

Co-authored-by: Zachary Hamm <zack@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>