fix(si): Fix the UX of secret update
The secret update command tried to find the secret by name but we will have multiple versions of a secret. Now a user will pass the component name and we will get the secret specifically attached to that component

This means that for secret defining components e.g. AWS Credential. we can update the secret but if a user tries to update the secret for an EC2 Instance, there’s no secret attached so the call will error.

The call will now be:
```
si secret update my-component —use-local-profile
```

This PR also deprecates the template-generate tool from the mcp server! This tool has been removed as templates are no longer supported via the API. We return an error to the user that suggests using the CLI directly