Add audit logs reference doc
This change adds the audit logs reference doc, which continues to follow
the "Diátaxis" docs pattern. The audit trail doc has been replaced, but
some of its contents made it along for the journey to the new doc. In
the sidebar, the doc is near the bottom due to relative importance. This
doc only leverages the Web Application interface since the CLI, agent
and public API do not yet support audit logs.

Signed-off-by: Nick Gerace <nick@systeminit.com>