chore(ci): remove artifacts that are no longer being built in CI
This change, in effect, ports the manually skipped constants in the
`si-ci` program in the `si/merge-queue` pipeline. If container images
are desired for the backend services there should be much easier
mechanisms nowadays, so the remaining `Dockerfile`s and related files
are removed where they wouldn't be useful anymore.

Importantly there are some container images that should be published and
promoted but for a subset of target platforms--these targets now use the
`platform_targets` array to declare these default exemptions.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>