fix(grafana) Use correct labels and metrics in LayerCache Metrics dashboard
Some of the label usage was wrong ("pg" instead of "postgres"), and one
of the metrics was using the wrong name (missing "_total"), which was
causing missing data, and some misleading behavior.

This also make sure that all timeseries will show all data points at
that time in tooltips, instead of only the specific one under the
cursor.