Add create templates how-to
This change adds a new how-to for creating templates. The how-to is
CLI-driven and contains a walkthrough as well as some tips and tricks.

In addition to the core change, the how-tos are now alphabetically
ordered and all developer resources are now underneath a non-linked
category in the sidebar for ease of use. The cloud providers are no
longer collapsed by default too.

Signed-off-by: Nick Gerace <nick@systeminit.com>