Split AWS VPC rather than ordering the how-tos
This change splits the AWS VPC how-to hot path into two distinct
how-tos rather than having an assumed, guided path. Users should be able
to start at any point in the documentation that they would like to,
even if we do have recommendations in a multiple choice list. This
change adds a redirect to the AI agent how-to since there are now
multiple options.

Signed-off-by: Nick Gerace <nick@systeminit.com>